fix: move modelByPk into new file utility.js
